agent-backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.31.0
sqlmodel>=0.0.8
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
